+++ src/raspi/io_controller/comms/receiver.py
@@ -1,6 +1,7 @@
 from potentiometer.pot import Potentiometer
-from pneumatics.pressure import Pressure
 
-current_data : list[int]
-potentiometers : list[Potentiometer]
-pressure : Pressure
+from raspi.io_controller.pneumatics.pressure.pressure import Pressure
+
+current_data: list[int]
+potentiometers: list[Potentiometer]
+pressure: Pressure
