+++ src/raspi/io_controller/pneumatics/__init__.py
@@ -1,0 +1,3 @@
+import compressor as compressor_actions
+import pressure as pressure_actions
+import valve as valve_actions
