+++ src/raspi/io_controller/pneumatics/compressor/__init__.py
@@ -1,0 +1,1 @@
+import compressor as compressor_actions
