+++ src/raspi/io_controller/pneumatics/pressure/util.py
@@ -1,8 +1,7 @@
 from functools import wraps
+
 from gpiozero import DigitalInputDevice
-
 from utils.deviceMock import FakeInputDevice
-
 
 pressure_pins = dict()
 
@@ -97,46 +96,3 @@ def pressure_action(func: callable) -> callable:
         return func(valve, *args[1:], **kwargs)
 
     return wrapper
-
-
-@pressure_action
-def is_pressure_ok(device) -> bool:
-    """
-    Check if the pressure sensor is ok.
-
-    Args:
-        name (str): the name of the pressure sensor
-
-    Returns:
-        (bool) true if the pressure sensor is ok, false otherwise
-    """
-    print("is_pressure_ok", device)
-    return device.is_active
-
-
-@pressure_action
-def on_pressure_active(device, action: callable) -> None:
-    """
-    Add a new pressure sensor change event.
-
-    Args:
-        name (str): the name of the pressure sensor
-        action (callable): the action to perform when the pressure sensor changes
-    """
-    if isinstance(device, str):
-        return
-    device.when_activated = action
-
-
-@pressure_action
-def on_pressure_deactive(device, action: callable) -> None:
-    """
-    Add a new pressure sensor change event.
-
-    Args:
-        name (str): the name of the pressure sensor
-        action (callable): the action to perform when the pressure sensor changes
-    """
-    if isinstance(device, str):
-        return
-    device.when_deactivated = action
