+++ src/raspi/io_controller/pneumatics/valve/util.py
@@ -1,6 +1,6 @@
 from functools import wraps
-from gpiozero import DigitalOutputDevice
 
+from gpiozero import DigitalOutputDevice
 
 valve_pins = dict()
 
@@ -82,62 +82,3 @@ def valve_action(func: callable) -> callable:
         return func(valve, *args[1:], **kwargs)
 
     return wrapper
-
-
-@valve_action
-def turn_valve_on(valve) -> None:
-    """
-    Turn a valve on.
-
-    Args:
-        valve (DigitalOutputDevice): the valve to turn on
-    """
-    valve.on()
-
-
-@valve_action
-def turn_valve_off(valve) -> None:
-    """
-    Turn a valve off.
-
-    Args:
-        valve (DigitalOutputDevice): the valve to turn off
-    """
-    valve.off()
-
-
-@valve_action
-def turn_valve(valve, state: bool) -> None:
-    """
-    Turn a valve on or off.
-
-    Args:
-        valve (DigitalOutputDevice): the valve to turn on or off
-        state (bool): `True` to turn the valve on, `False` to turn it off
-    """
-    turn_valve_on(valve) if state else turn_valve_off(valve)
-
-
-@valve_action
-def toggle_valve(valve) -> None:
-    """
-    Toggle a valve.
-
-    Args:
-        valve (DigitalOutputDevice): the valve to toggle
-    """
-    turn_valve(valve, not get_valve_state(valve))
-
-
-@valve_action
-def get_valve_state(valve) -> bool:
-    """
-    Get the state of a valve.
-
-    Args:
-        valve: the valve to get the state of
-
-    Returns:
-        (bool) `True` if the valve is on, `False` if it is off
-    """
-    return valve.value == 1
